# Improve Action View `translate` helper